Move scripts outside of lib folder
