+++ src/generateDiff.ts
@@ -1,9 +1,9 @@
 import fs from "fs";
 import path from "path";
-import { z } from "zod";
+import {z} from "zod";
 
-import {executeCommand, getDiffPath} from "@/lib/fileUtils";
-import {getFeaturesString} from "@/lib/utils";
+import {executeCommand, getDiffPath} from "@/fileUtils";
+import {getFeaturesString} from "@/utils";
 
 export const paramsSchema = z.object({
   currentVersion: z.string(),
@@ -19,18 +19,17 @@ export const paramsSchema = z.object({
 type Params = z.infer<typeof paramsSchema>;
 
 export default async function generateDiff(params: Params) {
-  const { success } = paramsSchema.safeParse(params);
+  const {success} = paramsSchema.safeParse(params);
   if (!success) {
-    return { error: "Invalid request body" };
+    return {error: "Invalid request body"};
   }
-  const { currentVersion, upgradeVersion, features } =
-    paramsSchema.parse(params);
+  const {currentVersion, upgradeVersion, features} = paramsSchema.parse(params);
   const featureFlags = Object.entries(features)
     .filter(([, value]) => value)
     .map(([key]) => `--${key}=true`)
     .join(" ");
 
-  const diffPath = getDiffPath({ currentVersion, upgradeVersion, features });
+  const diffPath = getDiffPath({currentVersion, upgradeVersion, features});
   const featuresString = getFeaturesString(features);
   const diffDir = `/tmp/${currentVersion}..${upgradeVersion}${
     featuresString ? `-${featuresString}` : ""
@@ -58,7 +57,7 @@ export default async function generateDiff(params: Params) {
   if (fs.existsSync(diffPath)) {
     const differences = fs.readFileSync(diffPath, "utf8");
 
-    return { differences, url };
+    return {differences, url};
   }
 
   try {
@@ -93,8 +92,8 @@ export default async function generateDiff(params: Params) {
     await executeCommand(`rm -rf ${diffDir}`);
 
     // Send the diff back to the client
-    return { differences, url };
+    return {differences, url};
   } catch (error) {
-    return { error };
+    return {error};
   }
 }
